// src/types/models/eft/common/tables/IQuest.ts
export interface IQuestRewards {
    AvailableForStart?: IQuestReward[];
    AvailableForFinish?: IQuestReward[];
    Started?: IQuestReward[];
    Success?: IQuestReward[];
    Fail?: IQuestReward[];
    FailRestartable?: IQuestReward[];
    Expired?: IQuestReward[];
}
export interface IQuestReward {
    value?: string | number;
    id?: string;
    type: QuestRewardType;
    index: number;
    target?: string;
    items?: Item[];
    loyaltyLevel?: number;
    /** Hideout area id */
    traderId?: string;
    unknown?: boolean;
    findInRaid?: boolean;
}